+++ R/infrastructure.R
@@ -453,31 +453,6 @@ use_build_ignore <- function(files, escape = TRUE, pkg = ".") {
   invisible(TRUE)
 }
 
-#' Add a git hook.
-#'
-#' @param hook Hook name. One of "pre-commit", "prepare-commit-msg",
-#'   "commit-msg", "post-commit", "applypatch-msg", "pre-applypatch",
-#'   "post-applypatch", "pre-rebase", "post-rewrite", "post-checkout",
-#'   "post-merge", "pre-push", "pre-auto-gc".
-#' @param script Text of script to run
-#' @param pkg package description, can be path or package name.  See
-#'   \code{\link{as.package}} for more information
-#' @export
-#' @family infrastructure
-#' @keywords internal
-use_git_hook <- function(hook, script, pkg = ".") {
-  pkg <- as.package(pkg)
-
-  hook_dir <- file.path(pkg$path, ".git", "hooks")
-  if (!file.exists(hook_dir)) {
-    stop("This project doesn't use git", call. = FALSE)
-  }
-
-  hook_path <- file.path(hook_dir, hook)
-  writeLines(script, hook_path)
-  Sys.chmod(hook_path, "0744")
-}
-
 #' Use README.Rmd
 #'
 #' This creates `README.Rmd` from template and adds to \code{.Rbuildignore}.
@@ -587,15 +562,6 @@ use_code_of_conduct <- function(pkg = ".") {
 
 add_build_ignore <- function(pkg = ".", files, escape = TRUE) {
   use_build_ignore(files, escape = escape, pkg = pkg)
-}
-
-add_git_ignore <- function(pkg = ".", ignores) {
-  pkg <- as.package(pkg)
-
-  path <- file.path(pkg$path, ".gitignore")
-  union_write(path, ignores)
-
-  invisible(TRUE)
 }
 
 union_write <- function(path, new_lines) {